refactor(protocol): split protocol modules
The protocol crate exposed every frame, mux, identity, image cache, and transport type from one shallow top-level Interface. That made callers learn unrelated concepts and made frame or transport bugs harder to localize.

Split the implementation into focused Modules, move tests to the Interface that owns each behavior, and update workspace imports to cross the narrow seams directly. This preserves behavior while improving Locality for frame materialization and transport framing.
