+++ crates/seance-render-test/tests/layer1_logic.rs
@@ -5,7 +5,7 @@
 //! behavior; a failure points at the specific invariant that regressed.
 
 use seance_mux::{LinkDetector, LinkModifiers, LinkSource, LinkTarget};
-use seance_protocol::GridPos;
+use seance_protocol::frame::GridPos;
 use seance_vt::test_support::HeadlessTerminal;
 
 #[test]
