+++ crates/seance-vt/src/frame.rs
@@ -1,6 +1,7 @@
 pub use seance_frame::{
     CellView, CellVisitor, FrameSource, ImageInfo, ImageVisitor, PlacementLayer, PlacementVisitor,
 };
-pub use seance_protocol::{
-    CellAttrs, CellColor, CursorInfo, CursorShape, DirtySnapshot, ImageId, PlacementSnapshot,
+pub use seance_protocol::frame::{
+    CellAttrs, CellColor, CursorInfo, CursorShape, DirtySnapshot, PlacementSnapshot,
 };
+pub use seance_protocol::identity::ImageId;
