+++ crates/seance-vt/src/snapshot.rs
@@ -1,2 +1,2 @@
 pub use seance_frame::SnapshotFrameSource;
-pub use seance_protocol::{RowMeta, SnapshotCell, SnapshotImage, VtSnapshot};
+pub use seance_protocol::frame::{RowMeta, SnapshotCell, SnapshotImage, VtSnapshot};
